Remove short_pwd function call
The function is in Zim's `prompt` module. We want to keep modules
independent from each other. In this case, themes should not be required
to depend on the `prompt` module.

The `short_pwd` functionality is actually completely removed, as this
prompt already makes paths shorter by making them relative to the git
root when in a git repo. Original shashankmehta's code didn't abbreviate
paths either. Also, there's enough horizontal space, as prompt does not
show much else, besides a couple of one-character indicators and the git
branch/commit.